+++ ZeefvanEratosthenes/Priemgetallen.cs
@@ -8,11 +8,11 @@ namespace ZeefvanEratosthenes
 {
     public class Priemgetallen
     {
-        public List<double> Zoek(int maximum)
+        public List<int> Zoek(int maximum)
         {
-            var getallen = new List<double>();
-            var output = new List<double>();
-            for (double i = 2; i <= maximum; i++)
+            var getallen = new List<int>();
+            var output = new List<int>();
+            for (int i = 2; i <= maximum; i++)
             {
                 getallen.Add(i);
             }
@@ -20,7 +20,7 @@ namespace ZeefvanEratosthenes
             {
                 var kleinsteGetal = getallen.OrderBy(x => x).FirstOrDefault();
                 output.Add(kleinsteGetal);
-                for (double i = 1; i <= maximum; i++)
+                for (int i = 1; i <= maximum; i++)
                 {
                     var veelvouden = kleinsteGetal * i;
                     getallen.Remove(veelvouden);
